+++ lib/screens/chat_invite_screen.dart
@@ -83,9 +83,9 @@ class ChatInviteScreen extends HookConsumerWidget {
     void handleAvatarTap() {
       final otherPubkey = chatProfile.data?.otherMemberPubkey;
       if (otherPubkey != null) {
-        unawaited(Routes.pushToChatInfo(context, otherPubkey));
+        unawaited(Routes.pushToInviteInfo(context, otherPubkey));
       } else {
-        Routes.pushToWip(context);
+        Routes.pushToGroupInfo(context, mlsGroupId);
       }
     }